requirements.txt
langchain==0.1.16
langchain-community==0.0.36
langchain-openai==0.1.6

openai==1.26.0
tiktoken==0.6.0

faiss-cpu==1.7.4

python-dotenv==1.0.1
redis==5.0.3

pypdf==4.2.0
scikit-learn==1.4.1.post1

sentence-transformers==2.6.1

fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
